+++ test/PerformanceTest.cpp
@@ -23,15 +23,15 @@ int main(int argc, char **argv) {
   MQProcessor::Queue<int, int, BufferSize> mqproc;
   std::vector<TestConsumer> consumers(countCons, TestConsumer());
   std::vector<Providers<BufferSize>> providers;
-  for (int i = 0; i < countCons; i++)
+  for (size_t i = 0; i < countCons; i++)
     mqproc.Subscribe(i, &consumers[i]);
-  for (int i = 0; i < countCons; i++)
+  for (size_t i = 0; i < countCons; i++)
     providers.emplace_back(Providers<BufferSize>(mqproc, countThreads, i));
   for (auto &i : providers)
     i.Start();
   for (auto &i : providers)
     i.Join();
-  for (int i = 0; i < countCons; i++) {
+  for (size_t i = 0; i < countCons; i++) {
     std::cout << providers[i].Sended() << ", " << consumers[i].consumed << '\n';
   }
   return 0;
